Moved createGridGraph to examples.Graphs.  Changed default sql shuffle partitions to 4 instead of 200 for tests.
