+++ src/main/scala/org/graphframes/examples/BeliefPropagation.scala
@@ -20,9 +20,10 @@ package org.graphframes.examples
 import org.apache.spark.{SparkConf, SparkContext}
 import org.apache.spark.graphx.{Edge => GXEdge,VertexRDD, Graph}
 import org.apache.spark.sql.{Row, SQLContext}
-import org.apache.spark.sql.functions.{col, udf, randn}
+import org.apache.spark.sql.functions.{col, udf}
 
 import org.graphframes.GraphFrame
+import org.graphframes.examples.Graphs.gridIsingModel
 
 
 /**
@@ -73,7 +74,7 @@ object BeliefPropagation {
 
     // Create graphical model g.
     val n = 3
-    val g = createGridGraph(sql, n)
+    val g = gridIsingModel(sql, n)
 
     // Run BP for 5 iterations.
     val results = runBP(g, 5)
@@ -86,71 +87,13 @@ object BeliefPropagation {
   }
 
   /**
-   * This method generates a grid Ising model with random parameters.
-   *
-   * Each vertex is parameterized by a single scalar a,,i,,.
-   * Each edge is parameterized by a single scalar b,,ij,,.
-   *
-   * @param  n  Length of one side of the grid.  The grid will be of size n x n.
-   * @param  vStd  Standard deviation of normal distribution used to generate vertex factors "a".
-   *               Default of 1.0.
-   * @param  eStd  Standard deviation of normal distribution used to generate edge factors "b".
-   *               Default of 1.0.
-   * @return  GraphFrame.  Vertices have columns "id" and "a".
-   *          Edges have columns "src", "dst", and "b".
-   *          Vertex IDs are of the form "i,j".  E.g., vertex "1,3" is in the second row and fourth
-   *          column of the grid.
-   */
-  def createGridGraph(sqlContext: SQLContext, n: Int, vStd: Double, eStd: Double): GraphFrame = {
-    require(n >= 1, s"Grid graph must have size >= 1, but was given invalid value n = $n")
-    val rows = sqlContext.createDataFrame(Range(0, n).map(i => Tuple1(i))).toDF("i")
-    val cols = rows.select(rows("i").as("j"))
-    // Cartesian join to create grid
-    val coordinates = rows.join(cols)
-
-    // Create SQL expression for converting coordinates (i,j) to a string ID "i,j"
-    val toIDudf = udf { (i: Int, j: Int) => i.toString + "," + j.toString }
-
-    // Create the vertex DataFrame
-    //  Create SQL expression for converting coordinates (i,j) to a string ID "i,j"
-    val vIDcol = toIDudf(col("i"), col("j"))
-    //  Add random parameters generated from a normal distribution
-    val seed = 12345
-    val vertices = coordinates.withColumn("id", vIDcol)  // vertex IDs "i,j"
-      .withColumn("a", randn(seed) * vStd)  // Ising parameter for vertex
-
-    // Create the edge DataFrame
-    //  Create SQL expression for converting coordinates (i,j+1) and (i+1.j) to string IDs
-    val rightIDcol = toIDudf(col("i"), col("j") + 1)
-    val downIDcol = toIDudf(col("i") + 1, col("j"))
-    val horizontalEdges = coordinates.filter(col("j") !== n - 1)
-      .select(vIDcol.as("src"), rightIDcol.as("dst"))
-    val verticalEdges = coordinates.filter(col("i") !== n - 1)
-      .select(vIDcol.as("src"), downIDcol.as("dst"))
-    val hvEdges = horizontalEdges.unionAll(verticalEdges)
-    //  These edges are directed.  We duplicate and reverse them to create bidirectional edges.
-    val reversedEdges = hvEdges.select(col("src").as("dst"), col("dst").as("src"))
-    //  We union all edges together to get the full set.
-    val allEdges = hvEdges.unionAll(reversedEdges)
-    //  Add random parameters from a normal distribution
-    val edges = allEdges.withColumn("b", randn(seed + 1) * eStd)  // Ising parameter for edge
-
-    // Create the GraphFrame
-    GraphFrame(vertices, edges)
-  }
-
-  /** Version of [[createGridGraph()]] with vStd, eStd set to 1.0. */
-  def createGridGraph(sqlContext: SQLContext, n: Int): GraphFrame =
-    createGridGraph(sqlContext, n, 1.0, 1.0)
-
-  /**
    * Given a GraphFrame, choose colors for each vertex.  No neighboring vertices will share the
    * same color.  The number of colors is minimized.
    *
    * This is written specifically for grid graphs. For non-grid graphs, it should be generalized,
    * such as by using a greedy coloring scheme.
    *
-   * @param g  Grid graph generated by [[createGridGraph()]]
+   * @param g  Grid graph generated by [[org.graphframes.examples.Graphs.gridIsingModel()]]
    * @return  Same graph, but with a new vertex column "color" of type Int (0 or 1)
    */
   private def colorGraph(g: GraphFrame): GraphFrame = {
@@ -169,7 +112,7 @@ object BeliefPropagation {
    *  - Run BP using GraphX's aggregateMessages API.
    *  - Augment the original GraphFrame with the BP results (vertex beliefs).
    *
-   * @param g  Graphical model created by [[createGridGraph()]]
+   * @param g  Graphical model created by [[org.graphframes.examples.Graphs.gridIsingModel()]]
    * @param numIter  Number of iterations of BP to run.  One iteration includes updating each
    *                 vertex's belief once.
    * @return  Same graphical model, but with [[GraphFrame.vertices]] augmented with a new column
