+++ src/test/scala/org/graphframes/GraphFrameTestSparkContext.scala
@@ -31,6 +31,7 @@ trait GraphFrameTestSparkContext extends BeforeAndAfterAll { self: Suite =>
     val conf = new SparkConf()
       .setMaster("local[2]")
       .setAppName("GraphFramesUnitTest")
+      .set("spark.sql.shuffle.partitions", "4")
     sc = new SparkContext(conf)
     sqlContext = new SQLContext(sc)
   }
